+++ end-to-end-tests/package-fs-mapping.sh
@@ -2,7 +2,7 @@
 
 alias wapm=target/debug/wapm
 wapm config set registry.url "https://registry.wapm.dev"
-yes 2> /dev/null | wapm install -g mark2/dog2@0.0.13
+wapm install -g mark2/dog2@0.0.13 --force-yes
 wapm run dog -- data
 wapm uninstall -g mark2/dog2
 wapm install mark2/dog2@0.0.13